+++ src/main/java/com/sellertool/auth_server/domain/signup/dto/SignupDto.java
@@ -24,7 +24,6 @@ public class SignupDto {
     @NotBlank
     private String nickname;
 
-    @NotBlank
     private String email;
 
     @NotBlank
