+++ src/main/java/com/sellertool/auth_server/domain/user/controller/UserApiV1.java
@@ -8,7 +8,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
@@ -28,12 +27,12 @@ public class UserApiV1 {
     }
 
     @GetMapping("/info/own")
-    public ResponseEntity<?> getInfoOwn(){
+    public ResponseEntity<?> getInfoOwn(HttpServletResponse response){
         Message message = new Message();
 
         message.setStatus(HttpStatus.OK);
         message.setMessage("success");
-        message.setData(userBusinessService.getInfoOwn());
+        message.setData(userBusinessService.getInfoOwn(response));
 
         return new ResponseEntity<>(message, message.getStatus());
     }
@@ -67,50 +66,6 @@ public class UserApiV1 {
         message.setStatus(HttpStatus.OK);
         message.setMessage("success");
         userBusinessService.changePassword(password);
-
-        return new ResponseEntity<>(message, message.getStatus());
-    }
-
-    @GetMapping("/phone")
-    public ResponseEntity<?> getPhoneAuthNumber(@RequestParam Map<String, Object> params, HttpServletResponse response){
-        Message message = new Message();
-
-        message.setStatus(HttpStatus.OK);
-        message.setMessage("success");
-        userBusinessService.getPhoneAuthNumber(params, response);
-
-        return new ResponseEntity<>(message, message.getStatus());
-    }
-
-    @GetMapping("/phone/verify")
-    public ResponseEntity<?> verifyPhoneAuthNumber(HttpServletRequest request, HttpServletResponse response, @RequestParam Map<String, Object> params){
-        Message message = new Message();
-
-        message.setStatus(HttpStatus.OK);
-        message.setMessage("success");
-        userBusinessService.verifyPhoneAuthNumber(request, response, params);
-
-        return new ResponseEntity<>(message, message.getStatus());
-    }
-
-    @GetMapping("/email")
-    public ResponseEntity<?> getEmailAuthNumber(@RequestParam Map<String, Object> params, HttpServletResponse response) throws IOException {
-        Message message = new Message();
-
-        message.setStatus(HttpStatus.OK);
-        message.setMessage("success");
-        userBusinessService.getEmailAuthNumber(params, response);
-
-        return new ResponseEntity<>(message, message.getStatus());
-    }
-
-    @GetMapping("/email/verify")
-    public ResponseEntity<?> verifyEmailAuthNumber(HttpServletRequest request, HttpServletResponse response, @RequestParam Map<String, Object> params){
-        Message message = new Message();
-
-        message.setStatus(HttpStatus.OK);
-        message.setMessage("success");
-        userBusinessService.verifyEmailAuthNumber(request, response, params);
 
         return new ResponseEntity<>(message, message.getStatus());
     }
